% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedSystematicError.R
\name{computeExpectedSystematicError}
\alias{computeExpectedSystematicError}
\title{Compute the expected systematic error}
\usage{
computeExpectedSystematicError(null)
}
\arguments{
\item{null}{An object of class \code{null} created using the \code{fitNull} function or an
object of class \code{mcmcNull} created using the \code{fitMcmcNull} function.}

\item{...}{Any additional parameters (currently none).}
}
\value{
The two-sided calibrated p-value.
}
\description{
\code{calibrateP} computes calibrated p-values using the fitted null distribution
}
\details{
This function computes a calibrated two-sided p-value as described in Schuemie et al (2014).
}
\examples{
data(sccs)
negatives <- sccs[sccs$groundTruth == 0, ]
null <- fitNull(negatives$logRr, negatives$seLogRr)
computeExpectedSystematicError(null)

}
