% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EmpiricalCalibration.R
\name{computeTraditionalP}
\alias{computeTraditionalP}
\title{Compute the (traditional) p-value}
\usage{
computeTraditionalP(logRr, seLogRr)
}
\arguments{
\item{logRr}{A numeric vector of one or more effect estimates on the log scale}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
                 error = (log(<lower bound 95 percent confidence interval>) - log(<effect
                 estimate>))/qnorm(0.025)}
}
\value{
A two-sided (traditional) p-value.
}
\description{
\code{computeTraditionalP} computes the traditional two-sided p-value based on the log of the
relative risk and the standerd error of the log of the relative risk.
}
\examples{
data(sccs)
positive <- sccs[sccs$groundTruth == 1, ]
computeTraditionalP(positive$logRr, positive$seLogRr)
}

