% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalCalibration.R
\docType{package}
\name{EmpiricalCalibration-package}
\alias{EmpiricalCalibration}
\alias{EmpiricalCalibration-package}
\title{EmpiricalCalibration: Routines for Performing Empirical Calibration of Observational Study Estimates}
\description{
Routines for performing empirical calibration of observational
  study estimates. By using a set of negative control hypotheses we can
  estimate the empirical null distribution of a particular observational
  study setup. This empirical null distribution can be used to compute a
  calibrated p-value, which reflects the probability of observing an
  estimated effect size when the null hypothesis is true taking both random
  and systematic error into account. A similar approach can be used to
  calibrate confidence intervals, using both negative and positive controls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/EmpiricalCalibration}
  \item \url{https://github.com/OHDSI/EmpiricalCalibration}
  \item Report bugs at \url{https://github.com/OHDSI/EmpiricalCalibration/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
}

}
\keyword{internal}
