% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCalibration}
\alias{plotCalibration}
\title{Create a calibration plot}
\usage{
plotCalibration(logRr, seLogRr)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025)}
}
\value{
A Ggplot object. Use the \code{ggsave} function to save to file.
}
\description{
\code{plotCalibration} creates a plot showing the calibration of our calibration procedure
}
\details{
Creates a calibration plot showing the number of effects with p < alpha for every level of alpha.
The empirical calibration is performed using a leave-one-out design: The p-value of an effect is computed
by fitting a null using all other negative controls. Ideally, the calibration line should approximate the
diagonal. The plot shows both theoretical (traditional) and empirically calibrated p-values.
}
\examples{
data(sccs)
negatives <- sccs[sccs$groundTruth == 0,]
plotCalibration(negatives$logRr,negatives$seLogRr)
}

