% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotForest}
\alias{plotForest}
\title{Create a forest plot}
\usage{
plotForest(logRr, seLogRr, names, xLabel = "Relative risk")
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
error = (log(<lower bound 95 percent confidence interval>) - log(<effect estimate>))/qnorm(0.025)}

\item{names}{A vector containing the names of the drugs or outcomes}

\item{xLabel}{The label on the x-axis: the name of the effect estimate}
}
\value{
A Ggplot object. Use the \code{ggsave} function to save to file.
}
\description{
\code{plotForest} creates a forest plot of effect size estimates.
}
\details{
Creates a forest plot of effect size estimates (ratios). Estimates that are significantly
different from 1 (alpha = 0.05) are marked in orange, others are marked in blue.
}
\examples{
data(sccs)
negatives <- sccs[sccs$groundTruth == 0,]
plotForest(negatives$logRr,negatives$seLogRr, negatives$drugName)
}

