% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalCalibration.R
\docType{package}
\name{EmpiricalCalibration-package}
\alias{EmpiricalCalibration}
\alias{EmpiricalCalibration-package}
\title{EmpiricalCalibration: Routines for Performing Empirical Calibration of Observational Study Estimates}
\description{
Routines for performing empirical calibration of observational study estimates. By using a set of negative control hypotheses we can estimate the empirical null distribution of a particular observational study setup. This empirical null distribution can be used to compute a calibrated p-value, which reflects the probability of observing an estimated effect size when the null hypothesis is true taking both random and systematic error into account. A similar approach can be used to calibrate confidence intervals, using both negative and positive controls. For more details, see Schuemie et al. (2013) \doi{10.1002/sim.5925} and Schuemie et al. (2018) \doi{10.1073/pnas.1708282114}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/EmpiricalCalibration/}
  \item \url{https://github.com/OHDSI/EmpiricalCalibration}
  \item Report bugs at \url{https://github.com/OHDSI/EmpiricalCalibration/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org} (\href{https://orcid.org/0000-0002-0817-5361}{ORCID})

Authors:
\itemize{
  \item Marc Suchard (\href{https://orcid.org/0000-0001-9818-479X}{ORCID})
}

}
\keyword{internal}
